# src/main/resources/static/api-docs.yml
swagger: "2.0"
info:
  description: >-
    Especificaciones técnicas para consumo de la api DigitalBanking
  version: 1.0.1
  title: Rest Api DigitalBanking
  termsOfService: http://swagger.io/terms/
  contact:
    email: soporte@digitalbanking.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
host: localhost:8080
basePath: /api/v1
...
schemes:
  - https
  - http
paths:
...
  /producto/{id}:
    get:
      tags:
        - Productos
      summary: Obtiene un producto por id
      description: ""
      operationId: buscarProductoPorId
      produces:
        - application/json
      parameters:
        - in: path
          type: string
          required: true
          name: id
          description: id
      responses:
        "200":
          description: Operacion exitosa
          schema:
            $ref: "#/definitions/ProductoDto"
        "400":
          description: Request incorrecto
          schema:
            $ref: "#/definitions/Error"
        "401":
          description: No autorizado
          schema:
            $ref: "#/definitions/Error"
        "403":
          description: Acceso prohibido
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: Error inesperado
          schema:
            $ref: "#/definitions/Error"
      deprecated: false
    delete:
      tags:
        - Productos
      summary: Elimina un producto
      description: ""
      operationId: eliminarProducto
      produces:
        - application/json
      parameters:
        - in: path
          type: string
          name: id
          required: true
          description: id
      responses:
        "200":
          description: Operacion exitosa
        "400":
          description: Request incorrecto
          schema:
            $ref: "#/definitions/Error"
        "401":
          description: No autorizado
          schema:
            $ref: "#/definitions/Error"
        "403":
          description: Acceso prohibido
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: Error inesperado
          schema:
            $ref: "#/definitions/Error"
      security:
        - APIKeyHeader: [ ]
      deprecated: false
  /producto/sede/{id}:
    post:
      tags:
        - Productos
      summary: Obtiene los productos disponibles en una sede
      description: ""
      operationId: buscarProductosPorSede
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          description: id de la sede
          type: integer
          required: true
        - in: body
          name: request
          description: request
          schema:
            $ref: "#/definitions/PagineableDto"
      responses:
        "200":
          description: Operacion exitosa
          schema:
            type: array
            items:
              $ref: "#/definitions/ProductoDto"
        "400":
          description: Request incorrecto
          schema:
            $ref: "#/definitions/Error"
        "401":
          description: No autorizado
          schema:
            $ref: "#/definitions/Error"
        "403":
          description: Acceso prohibido
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: Error inesperado
          schema:
            $ref: "#/definitions/Error"
      deprecated: false
...
  /rol/buscarTodos:
    post:
      tags:
        - Roles
      summary: Obtiene todos las roles
      description: ""
      operationId: buscarRoles
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: request
          schema:
            $ref: "#/definitions/PagineableDto"
      responses:
        "200":
          description: Operacion exitosa
          schema:
            type: array
            items:
              $ref: "#/definitions/RolDto"
        "400":
          description: Request incorrecto
          schema:
            $ref: "#/definitions/Error"
        "401":
          description: No autorizado
          schema:
            $ref: "#/definitions/Error"
        "403":
          description: Acceso prohibido
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: Error inesperado
          schema:
            $ref: "#/definitions/Error"
      security:
        - APIKeyHeader: [ ]
      deprecated: false
...
  /rol/{id}:
    get:
      tags:
        - Roles
      summary: Obtiene un rol por id
      description: ""
      operationId: buscarRolPorId
      produces:
        - application/json
      parameters:
        - in: path
          type: string
          required: true
          name: id
          description: id
      responses:
        "200":
          description: Operacion exitosa
          schema:
            $ref: "#/definitions/RolDto"
        "400":
          description: Request incorrecto
          schema:
            $ref: "#/definitions/Error"
        "401":
          description: No autorizado
          schema:
            $ref: "#/definitions/Error"
        "403":
          description: Acceso prohibido
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: Error inesperado
          schema:
            $ref: "#/definitions/Error"
      security:
        - APIKeyHeader: [ ]
      deprecated: false
    delete:
      tags:
        - Roles
      summary: Elimina un rol
      description: ""
      operationId: eliminarRol
      produces:
        - application/json
      parameters:
        - in: path
          type: string
          name: id
          required: true
          description: id
      responses:
        "200":
          description: Operacion exitosa
        "400":
          description: Request incorrecto
          schema:
            $ref: "#/definitions/Error"
        "401":
          description: No autorizado
          schema:
            $ref: "#/definitions/Error"
        "403":
          description: Acceso prohibido
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: Error inesperado
          schema:
            $ref: "#/definitions/Error"
      security:
        - APIKeyHeader: [ ]
      deprecated: false
  /rol:
    post:
      tags:
        - Roles
      summary: Crea un rol nuevo
      description: ""
      operationId: crearRol
      produces:
        - application/json
      parameters:
        - in: body
          name: id
          description: request
          schema:
            $ref: "#/definitions/CategoriaDto"
      responses:
        "200":
          description: Operacion exitosa
          schema:
            $ref: "#/definitions/RolDto"
        "400":
          description: Request incorrecto
          schema:
            $ref: "#/definitions/Error"
        "401":
          description: No autorizado
          schema:
            $ref: "#/definitions/Error"
        "403":
          description: Acceso prohibido
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: Error inesperado
          schema:
            $ref: "#/definitions/Error"
      security:
        - APIKeyHeader: [ ]
      deprecated: false
    put:
      tags:
        - Roles
      summary: Actualizar un rol
      description: ""
      operationId: actualizarRol
      produces:
        - application/json
      parameters:
        - in: body
          name: id
          description: request
          schema:
            $ref: "#/definitions/CategoriaDto"
      responses:
        "200":
          description: Operacion exitosa
          schema:
            $ref: "#/definitions/CategoriaDto"
        "400":
          description: Request incorrecto
          schema:
            $ref: "#/definitions/Error"
        "401":
          description: No autorizado
          schema:
            $ref: "#/definitions/Error"
        "403":
          description: Acceso prohibido
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: Error inesperado
          schema:
            $ref: "#/definitions/Error"
      security:
        - APIKeyHeader: [ ]
      deprecated: false
...
  /usuario/{id}:
    get:
      tags:
        - Usuarios
      summary: Obtiene un usuario por id
      description: ""
      operationId: buscarUsuarioPorId
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          description: id
          required: true
          type: integer
      responses:
        "200":
          description: Operacion exitosa
          schema:
            $ref: "#/definitions/UsuarioDto"
        "400":
          description: Request incorrecto
          schema:
            $ref: "#/definitions/Error"
        "401":
          description: No autorizado
          schema:
            $ref: "#/definitions/Error"
        "403":
          description: Acceso prohibido
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: Error inesperado
          schema:
            $ref: "#/definitions/Error"
      security:
        - APIKeyHeader: [ ]
      deprecated: false
    put:
      tags:
        - Usuarios
      summary: Actualiza un usuario
      description: ""
      operationId: actualizarUsuario
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          description: id
          required: true
          type: integer
        - in: body
          name: body
          description: request
          schema:
            $ref: "#/definitions/ActualizarUsuarioDto"
      responses:
        "200":
          description: Operacion exitosa
          schema:
            $ref: "#/definitions/UsuarioDto"
        "400":
          description: Request incorrecto
          schema:
            $ref: "#/definitions/Error"
        "401":
          description: No autorizado
          schema:
            $ref: "#/definitions/Error"
        "403":
          description: Acceso prohibido
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: Error inesperado
          schema:
            $ref: "#/definitions/Error"
      security:
        - APIKeyHeader: [ ]
      deprecated: false
...
securityDefinitions:
  APIKeyHeader:
    type: apiKey
    in: header
    name: Authorization
definitions:
  PagineableDto:
    type: object
    properties:
      pagina:
        type: integer
        format: int32
      cantidad:
        type: integer
        format: int32
  CategoriaDto:
    type: object
    properties:
      idCategoria:
        type: integer
        format: int32
      codigo:
        type: string
      nombre:
        type: string
      descripcion:
        type: string
      fechaCreacion:
        type: string
        format: datetime
      indicadorHabilitado:
        type: boolean
  ProductoDto:
    type: object
    properties:
      idProducto:
        type: integer
        format: int32
      idCategoria:
        type: integer
        format: int32
      nombre:
        type: string
      marca:
        type: string
      modelo:
        type: string
      precio:
        type: number
      descripcion:
        type: string
      imagen:
        type: string
      indicadorHabilitado:
        type: boolean
...
  RolDto:
    type: object
    properties:
      idRol:
        type: integer
        format: int32
      codigo:
        type: string
      nombre:
        type: string
      indicadorHabilitado:
        type: boolean
  UsuarioDto:
    type: object
    properties:
      idUsuario:
        type: integer
      usuario:
        type: string
      rol:
        type: object
        properties:
          nombre:
            type: string
      persona:
        type: object
        properties:
          nombre:
            type: string
          apellido:
            type: string
          correo:
            type: string
      indicadorHabilitado:
        type: boolean
      verificado:
        type: boolean
      tokenValidacionCorreo:
        type: string
      fechaExpiracionToken:
        type: string
        format: datetime
...
  ActualizarUsuarioDto:
    type: object
    properties:
      idUsuario:
        type: integer
      nombre:
        type: string
      apellido:
        type: string
      correo:
        type: string
      usuario:
        type: string
      contrasenia:
        type: string
      rol:
        type: integer
        format: int32
      indicadorHabilitado:
        type: boolean
      verificado:
        type: boolean
...
  Error:
    type: object
    properties:
      timestamp:
        type: integer
        format: int32
      status:
        type: integer
        format: int32
      error:
        type: string
      path:
        type: string
...